pyproject.toml
[project]
name = "deppenklopfer-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "email-validator>=2.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.11.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.38.0",
]
